requirements.txt
google-generativeai==0.8.3
langchain==0.3.14
langchain-google-genai==2.0.7
psycopg2-binary==2.9.10
pydantic==2.10.4
gradio==5.9.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
